ignite-5392 fix compile error after merge

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/CacheData.java
@@ -180,7 +180,7 @@ public class CacheData implements Serializable {
     /**
      * @return Flags.
      */
-    public byte flags() {
+    public long flags() {
         return flags;
     }
